+++ synapse-core/src/main/java/de/otto/synapse/configuration/SynapseAutoConfiguration.java
@@ -15,7 +15,7 @@ import org.springframework.context.annotation.Role;
 
 import java.util.Map;
 
-import static de.otto.synapse.translator.ObjectMappers.defaultObjectMapper;
+import static de.otto.synapse.translator.ObjectMappers.currentObjectMapper;
 import static org.slf4j.LoggerFactory.getLogger;
 import static org.springframework.beans.factory.config.BeanDefinition.ROLE_INFRASTRUCTURE;
 
@@ -29,7 +29,7 @@ public class SynapseAutoConfiguration {
 
     @Bean
     public ObjectMapper objectMapper() {
-        return defaultObjectMapper();
+        return currentObjectMapper();
     }
 
     @Bean