+++ synapse-core/src/main/java/de/otto/synapse/state/ChronicleMapStateRepository.java
@@ -9,7 +9,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Optional;
 
-import static de.otto.synapse.translator.ObjectMappers.defaultObjectMapper;
+import static de.otto.synapse.translator.ObjectMappers.currentObjectMapper;
 
 public class ChronicleMapStateRepository<V> extends StateRepository<V> {
 
@@ -61,7 +61,7 @@ public class ChronicleMapStateRepository<V> extends StateRepository<V> {
 
     public static final class Builder<V> {
 
-        private ObjectMapper objectMapper = defaultObjectMapper();
+        private ObjectMapper objectMapper = currentObjectMapper();
 
         private final Class<V> clazz;
         private ChronicleMapBuilder<String, V> chronicleMapBuilder;